+++ Modules/Intent.Modules.Rdbms.Importer/DesignerScripts/strategy-database-import.ts
@@ -206,7 +206,8 @@ class DatabaseImportStrategy {
                             selectOptions: [
                                 { id: "Default", description: "(Default)" },
                                 { id: "StoredProcedureElement", description: "Stored Procedure Element" },
-                                { id: "RepositoryOperation", description: "Stored Procedure Operation" }
+                                { id: "RepositoryOperation", description: "Stored Procedure Operation" },
+                                { id: "RepositoryOperationMapping", description: "Stored Procedure Operation mapped to Element" }
                             ]
                         }
                     ],