+++ Modules/Intent.Modules.Rdbms.Importer/resources/scripts/sql-importer-core.js
@@ -222,7 +222,8 @@ class DatabaseImportStrategy {
                             selectOptions: [
                                 { id: "Default", description: "(Default)" },
                                 { id: "StoredProcedureElement", description: "Stored Procedure Element" },
-                                { id: "RepositoryOperation", description: "Stored Procedure Operation" }
+                                { id: "RepositoryOperation", description: "Stored Procedure Operation" },
+                                { id: "RepositoryOperationMapping", description: "Stored Procedure Operation mapped to Element" }
                             ]
                         }
                     ],
@@ -899,7 +900,8 @@ class StoredProceduresImportStrategy {
                     selectOptions: [
                         { id: "", description: "(default or inherited setting)" },
                         { id: "StoredProcedureElement", description: "Stored Procedure Element" },
-                        { id: "RepositoryOperation", description: "Stored Procedure Operation" }
+                        { id: "RepositoryOperation", description: "Stored Procedure Operation" },
+                        { id: "RepositoryOperationMapping", description: "Stored Procedure Operation mapped to Element" }
                     ]
                 },
                 {